64-bit integration: - upgrade of widgetlibrary SDK
git-svn-id: https://svn.code.sf.net/p/x-ivap/code/trunk@82 c9bdb542-074d-4938-be84-19bf99c82884

+++ WidgetLibrary/SDK/CHeaders/Widgets/XPWidgetUtils.h
@@ -2,11 +2,11 @@
 #define _XPWidgetUtils_h_
 
 /*
- * Copyright 2005 Sandy Barbour and Ben Supnik
+ * Copyright 2005-2012 Sandy Barbour and Ben Supnik
  * 
  * All rights reserved.  See license.txt for usage.
  * 
- * X-Plane SDK Version: 1.0.2                                                  
+ * X-Plane SDK Version: 2.1.1                                                  
  *
  */
 
@@ -47,6 +47,7 @@ extern "C" {
  *                                                                             
  *
  */
+
 
 
 
@@ -125,7 +126,7 @@ typedef struct {
  */
 WIDGET_API void                 XPUCreateWidgets(
                                    const XPWidgetCreate_t * inWidgetDefs,    
-                                   long                 inCount,    
+                                   int                  inCount,    
                                    XPWidgetID           inParamParent,    
                                    XPWidgetID *         ioWidgets);    
 
@@ -151,6 +152,8 @@ WIDGET_API void                 XPUMoveWidgetBy(
  *
  */
 
+
+
 /*
  * XPUFixedLayout
  * 
@@ -162,8 +165,8 @@ WIDGET_API void                 XPUMoveWidgetBy(
 WIDGET_API int                  XPUFixedLayout(
                                    XPWidgetMessage      inMessage,    
                                    XPWidgetID           inWidget,    
-                                   long                 inParam1,    
-                                   long                 inParam2);    
+                                   intptr_t             inParam1,    
+                                   intptr_t             inParam2);    
 
 /***************************************************************************
  * WIDGET PROC BEHAVIORS
@@ -174,6 +177,8 @@ WIDGET_API int                  XPUFixedLayout(
  * your widget function.                                                       
  *
  */
+
+
 
 /*
  * XPUSelectIfNeeded
@@ -186,8 +191,8 @@ WIDGET_API int                  XPUFixedLayout(
 WIDGET_API int                  XPUSelectIfNeeded(
                                    XPWidgetMessage      inMessage,    
                                    XPWidgetID           inWidget,    
-                                   long                 inParam1,    
-                                   long                 inParam2,    
+                                   intptr_t             inParam1,    
+                                   intptr_t             inParam2,    
                                    int                  inEatClick);    
 
 /*
@@ -200,8 +205,8 @@ WIDGET_API int                  XPUSelectIfNeeded(
 WIDGET_API int                  XPUDefocusKeyboard(
                                    XPWidgetMessage      inMessage,    
                                    XPWidgetID           inWidget,    
-                                   long                 inParam1,    
-                                   long                 inParam2,    
+                                   intptr_t             inParam1,    
+                                   intptr_t             inParam2,    
                                    int                  inEatClick);    
 
 /*
@@ -215,8 +220,8 @@ WIDGET_API int                  XPUDefocusKeyboard(
 WIDGET_API int                  XPUDragWidget(
                                    XPWidgetMessage      inMessage,    
                                    XPWidgetID           inWidget,    
-                                   long                 inParam1,    
-                                   long                 inParam2,    
+                                   intptr_t             inParam1,    
+                                   intptr_t             inParam2,    
                                    int                  inLeft,    
                                    int                  inTop,    
                                    int                  inRight,
